+++ Classes/Phoenix.Classes.psm1
@@ -8250,6 +8250,7 @@ class PhoenixRestoreCheckpoint {
     [string]$Status
     [bool]$RebootRequired
     [string]$PhoenixVersion
+    [object]$PlanSnapshot
     [PhoenixRestoreCheckpointRecord[]]$Records
     [string]$StoragePath